newpkg: Complete rewrite to use a modular system.
newpkg will run any defined "presets"; default.preset if none is given.
Presets may define two things: copies and modifications. Common sets of
those are split into "actions" -- modular building blocks. The default
preset includes the 'pkgbuild' action, which defines one copy (The proto
PKGBUILD), and several modifications (setting pkgname, contributor,
stripping boilerplate comments, so on). This system allows for a very
flexible creation process which I hope to make even better in the
future. An interactive preset similar to debian's dh_make should be easy
to add.

Once the preset is sourced and any actions defined either in that preset
or by the user on the command line, newpkg does 3 things;
1) Make the pkgbuild directory
2) Copy all defined files into it
3) Run all defined modifications

For now, command line options are translated to their
equivalents from the old version, but they are not guaranteed to stay
the same.

New options added are -a (--action) and -P (--preset).