Return a dict with 'files' key being a list in search_file
Instead of returning a dict for every matching file,
collect all the matching files of a given package in a list,
and return this list in the 'files' key of the result dict.

In the result list, there is now only one dict per package.

Based on a patch by solsTiCe d'Hiver <solstice.dhiver@gmail.com>
but refactored a bit and with memleaks removed.