Import of gem2arch
This is a new tool that, given a Ruby gem name, will download and read
the gem's specification and intellegently generate a generic gem-based
PKGBUILD. Alternatively, gem2arch can also list a gem's dependencies
based on the same mechanism, allowing packages to be recursively built.

Included in this import are the relevant Makefile changes, .gitignore
changes, gem2arch manpage, and of course the source.

Signed-off-by: Andrei 'Garoth' Thorp <garoth@gmail.com>