PKGBUILD-git: Fix possible vercmp issue with pkgver() function
This inserts the number of commits between the date and the commit hash.

This should resolve an issue that could occur when building the package
twice in one day when multiple commits were made on the same day where
the latter commit(s)' hash would trick vercmp into thinking a package
downgrade was happening.